Fix route for About page
No Issue

- Routes,  views and controllers updated for about page
- fix up tests and add redirect check
