+++ core/client/app/templates/about.hbs
@@ -3,7 +3,7 @@
     <header class="gh-about-header">
         <img class="gh-logo" src="{{gh-path 'admin' '/img/ghost-logo.png'}}" alt="Ghost" />
         <!-- TODO: fix about notifications -->
-        {{gh-notifications location="settings-about-upgrade" notify="updateNotificationChange"}}
+        {{gh-notifications location="about-upgrade" notify="updateNotificationChange"}}
     </header>
 
     <section class="gh-env-details">
