+++ core/client/app/views/about.js
@@ -1,0 +1,5 @@
+import BaseView from 'ghost/views/settings/content-base';
+
+var AboutView = BaseView.extend();
+
+export default AboutView;
