--- core/client/app/views/settings/about.js
@@ -1,5 +1,0 @@
-import BaseView from 'ghost/views/settings/content-base';
-
-var SettingsAboutView = BaseView.extend();
-
-export default SettingsAboutView;
